Refactor AccountAuthorizationService to support generics.
Replaced the implementation of `AccountAuthorizationService` with a generic interface and introduced `IdAccountAuthorizationService` for account ID-based validation. Updated tests accordingly to use the new implementation, ensuring backward compatibility and improved flexibility for future extensions.
